Revert "second-step: main service"
This reverts commit e23e47c01591ebcbf365dac4cfea5dd40f5d9185.

+++ ewm-stats-server/stats-service/src/main/java/ru/practicum/explorewithme/statserver/model/StatWithHits.java
@@ -6,6 +6,6 @@ public interface StatWithHits {
 
     String getUri();
 
-    Integer getHits();
+    Long getHits();
 
 }
